Added basic support for history and better object copying.
A requirement variable can store a discrete history of its values, to be used in the checks.
The checker can use the ReqVar name, with [] appended to it.
This currently has a limitation that requirement variable names must be unique, even across theorems.

Annotated variables can now be any primitive type, or any object that has a copy constructor, as long
as that copy constructor performs a deep copy. Also, it can be an array of any of the other supported types.
Some common types, like Integer are not currently supported, so the corresponding primitive types
should be used.